moved ajax handlers to new file
